abstract down positional considerations (we should probably reconfigure some tests, but, they do the job for now)
